requirements.txt
anyio>=4.3.0
async-timeout>=4.0.3
certifi>=2024.2.2
cffi>=1.16.0
charset-normalizer>=3.3.2
click>=8.1.7
cryptography>=42.0.5
exceptiongroup>=1.2.0
fastapi>=0.110.0
h11>=0.14.0
idna>=3.6
pycparser>=2.21
pydantic>=2.6.3
pydantic_core>=2.16.3
python-dotenv>=1.0.1
redis>=5.0.2
requests>=2.31.0
sniffio>=1.3.1
starlette>=0.36.3
types-pyOpenSSL>=24.0.0.20240228
types-redis>=4.6.0.20240218
typing_extensions>=4.10.0
urllib3>=2.2.1
uvicorn>=0.27.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
docker==7.0.0
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
packaging==23.2
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
types-pyOpenSSL==24.0.0.20240228
types-redis==4.6.0.20240218
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
